Moved function get_cv2_image_size to separate module
